test(bdd): refine Run BDD and cleanup obsolete tests
